backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
python-dotenv
pydantic
pdfplumber
Pillow
python-multipart
imapclient
email-validator
alembic
pytest
requests
